# add special case for findlast on tuples of length >= 32, fixes #45117