--- a/base/tuple.jl
+++ b/base/tuple.jl
@@ -418,7 +418,12 @@ function _findfirst_loop(f::Function, t)
     return nothing
 end
 findfirst(f::Function, t::Tuple) = length(t) < 32 ? _findfirst_rec(f, 1, t) : _findfirst_loop(f, t)
-findlast(f::Function, t::Tuple) = length(t) < 32 ? findfirst(f, reverse(t)) : _findlast_loop(f, t)
+
+findlast(f::Function, t::Tuple) = length(t) < 32 ? _findlast_rec(f, t) : _findlast_loop(f, t)
+function _findlast_rec(f::Function, x::Tuple)
+    r = findfirst(f, reverse(x))
+    return isnothing(r) ? r : length(x) - r + 1
+end
 function _findlast_loop(f::Function, t)
     for i in reverse(1:length(t))
         f(t[i]) && return i