@KristofferC 2022-05-10
Just to check, are there already tests for tuples with both fewer and more elements than 32 for `findlast`?