@pcjentsch 2022-05-10
> Just to check, are there already tests for tuples with both fewer and more elements than 32 for `findlast`?

There aren't any for `findlast` or `findfirst` for long tuples. I have added some, and also fixed a mistake in my code.